Wait for all management tasks to complete before shutting down CLSM
When draining, we shut down the CommitLog instance, which in turn shuts down
the CommitLogSegmentManager. Shutting it down without waiting for all its
management tasks to complete can lead to scheduling commitlogs for deletion,
but not actually deleting them, leading to a longer recovery after a restart.

Change the CLSM shutdown method to wait for all currently scheduled tasks to
complete before returning.