CASSANDRA-13649: Ensure all exceptions are correctly handled in the netty pipeline
Motivation:

To ensure we handle all exceptions in the netty pipeline that are produced by either the Channel itself or the ChannelHandlers in the pipeline we need to ensure the handlers that does so not uses the RequestThreadPoolExecutor as it not enforces strict ordering of the events per Channel.

Modifications:

Add a special ChannelInboundHandler implementations that handles the exceptions and move the exceptionCaught(...) impl from Dispatcher to the new handler (ExceptionHandler).

Result:

All exceptions produced by the Channel or the ChannelHanders in the ChannelPipeline should be correctly handled.