Fix AbstractTokenTreeBuilder#serializedSize() when there is a single leaf and overflow collisions
* overflow collisions were not included in the leaf size calulation
when there is a single block. This didn't affect multi-block trees
because those trees always have complete blocks written. This makes
hitting this bug extremely unlikely: it requires >= 3 collisions in a
small tree.

* removed collidingTokensMap in TokenTreeTest because its no longer
used for the merge tests and isn't useful compared to the other two
defined maps in TokenTreeTest

* made tests cover bot remaining token maps in TokenTreeTest